web-animations/animation-player/animation-player-start-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer startTime attribute test</title>
<meta name="assert" content="Returns the start time of this player. Setting this attribute updates the player start time using the procedure to update the player start time of this object to the new value">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-starttime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = newAnimationPlayer(createDiv(this));

    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.startTime, newStartTime,
        'The value of AnimationPlayer.startTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.startTime attribute to new value');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime + 50;
        player.startTime = newStartTime;
        // New player start time is in 50ms.
        // Wait for 100ms and chek that currentTime is about 50ms
        setTimeout(function() {
            t.step(function () {
                assert_approx_equals(player.currentTime, 50, EPSILON, 'The value of ' +
                    'AnimationPlayer.currentTime should correspond to startTime');
                t.done();
            });
        }, 100);        
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time in the future delays moves beginning of play ' +
    'accordingly');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));

    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime - 100;
        player.startTime = newStartTime;
        setTimeout(function() {
            t.step(function () {
                assert_approx_equals(player.currentTime, 150, EPSILON, 'The value of ' +
                    'AnimationPlayer.currentTime should correspond to startTime');
                t.done();
            });
        }, 50);
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time in the past shifts beginning of play');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.ready.then(t.step_func(function() {
        var newStartTime = document.timeline.currentTime - ANIMATION_END_TIME - 100;
        player.startTime = newStartTime;
        assert_equals(player.currentTime, ANIMATION_END_TIME, 'The value of ' +
            'AnimationPlayer.currentTime should correspond updated startTime');
        assert_equals(player.playState, 'finished', 'Player should be finished');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test setting AnimationPlayer start time to the value in the past that ' +
    'earlier that current time minus animation length finishes the player');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    var newStartTime = document.timeline.currentTime + 100;
    player.startTime = newStartTime;
    player.ready.then(t.step_func(function(plr) {
        assert_equals(plr, player, 'The current ready promise should be fulfilled with the player');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'Test that if player has a pending play task player\'s current ready promise ' +
    'is fulfilled with player');

// TODO Add tests for setting null start time value and setting a value when 
// Animation player has no timeline associated
</script>
</body>
